1、用户管理模块，新增、修改、删除、重置密码 2、优化前端dict组件 3、密码处理方法 4、调整其它页面优化部分问题

+++ back-end/color-system/src/main/java/com/colorone/system/domain/entity/BaseUserRole.java
@@ -21,11 +21,11 @@ public class BaseUserRole extends BaseEntity {
     /**
      * comment '用户ID'  bigint(20) not null
      */
-    Integer userId;
+    Long userId;
 
     /**
      * comment '角色ID'  bigint(20) not null
      */
-    Integer roleId;
+    Long roleId;
 
 }
